Failing because not even text lines covered
Failed the achievement when equipping and item with additional abilities on the gear like "Equip: Improves Critical Strike Rating by 3". Because it increased the amount of regions it needed to cycle through. increased i >= 36 to 36 because that is the total number of regions returned by select("#", GameTooltip:GetRegions()), so we should never go over that number and we should cover all the lines of tooltip text.